docs: 没自报额度就一次一个，理由写进代码与 CONTEXT；额度类型统一成 u32
`unwrap_or(1)` 看着像一个保守的猜测，其实是那台 agent 的实情：#260 之前的 sink 是
`for request in server.incoming_requests()`，一个请求一个请求地处理，同一时刻只跑得动
一个 run，而不带 `max_concurrent_runs` 字段的 agent 恰恰就是那批。sink.toml 的默认值 4
是**配了新版本的人**选的数，借给一台从来没有过这个字段的 agent 才是猜。这段理由原来
只活在评审记录里，下一个人很容易顺手把它「修」成 4——所以写进取值那一行，也写进 CONTEXT。

顺带 T5：同一个概念两种类型。`AgentInfo::max_concurrent_runs` 是 `Option<usize>`，
source 注册表里是 `Option<u32>`，中间夹一次 `try_from(...).unwrap_or(MAX)`。报文里统一成
`u32`——宽度按机器变的类型不该出现在两端之间的报文里——那次转换随之消失。

+++ crates/shared/src/protocol.rs
@@ -356,8 +356,11 @@ pub struct AgentInfo {
     ///
     /// **旧版本 agent 不带这个字段**，所以是 `default` + `skip_serializing_if`；读到 `None`
     /// 的一方按「这台 agent 没说」处理，不许拿 4 顶上——见 `source` 侧调度器的取值规则。
+    ///
+    /// `u32` 而不是 `usize`：同一个概念在线上、在 source 的注册表里只有一种类型，
+    /// 而 `usize` 的宽度按机器变，不该出现在两端之间的报文里。
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_concurrent_runs: Option<usize>,
+    pub max_concurrent_runs: Option<u32>,
 }
 
 /// 目标端 MySQL 的自述：版本，以及生成建表语句要用的那一项字符序。
